+++ src/main/java/roomescape/presentation/reservation/ReservationController.java
@@ -2,7 +2,6 @@ package roomescape.presentation.reservation;
 
 import jakarta.validation.Valid;
 import java.net.URI;
-import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -10,62 +9,39 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.reservation.ReservationService;
-import roomescape.application.reservation.dto.request.AdminReservationRequest;
-import roomescape.application.reservation.dto.request.ReservationFilterRequest;
 import roomescape.application.reservation.dto.request.ReservationRequest;
 import roomescape.application.reservation.dto.response.ReservationResponse;
 import roomescape.auth.LoginMemberId;
 
 @RestController
+@RequestMapping("/reservations")
 public class ReservationController {
-    private final ReservationService service;
+    private final ReservationService reservationService;
 
-    public ReservationController(ReservationService service) {
-        this.service = service;
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<ReservationResponse>> findAll() {
-        List<ReservationResponse> responses = service.findAll();
+        List<ReservationResponse> responses = reservationService.findAll();
         return ResponseEntity.ok(responses);
     }
 
-    @PostMapping("/reservations")
+    @PostMapping
     public ResponseEntity<ReservationResponse> create(@LoginMemberId long memberId,
                                                       @RequestBody @Valid ReservationRequest request) {
-        ReservationResponse response = service.create(memberId, request);
+        ReservationResponse response = reservationService.create(memberId, request);
         URI location = URI.create("/reservations/" + response.id());
         return ResponseEntity.created(location).body(response);
     }
 
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> createByAdmin(@RequestBody @Valid AdminReservationRequest request) {
-        ReservationRequest reservationRequest = new ReservationRequest(
-                request.date(),
-                request.timeId(),
-                request.themeId()
-        );
-        ReservationResponse response = service.create(request.memberId(), reservationRequest);
-        URI location = URI.create("/reservations/" + response.id());
-        return ResponseEntity.created(location).body(response);
-    }
-
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable long id) {
-        service.deleteById(id);
+        reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/admin/reservations")
-    public ResponseEntity<List<ReservationResponse>> findByFilter(@RequestParam(required = false) Long memberId,
-                                                                  @RequestParam(required = false) Long themeId,
-                                                                  @RequestParam(required = false) LocalDate startDate,
-                                                                  @RequestParam(required = false) LocalDate endDate) {
-        ReservationFilterRequest request = new ReservationFilterRequest(memberId, themeId, startDate, endDate);
-        List<ReservationResponse> responses = service.findByFilter(request);
-        return ResponseEntity.ok(responses);
     }
 }
